+++ NEWS.md
@@ -1,4 +1,4 @@
-# v 0.2.14.00X
+# v 0.2.15
 
 ## Breaking changes: